# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 2026 Schools / National Check-in Week API
  description: |
    API documentation for the 2026 Schools platform - National Check-in Week student wellbeing initiative.
    
    ## Authentication
    All `/api/admin/*` endpoints require authentication using Bearer token in the Authorization header.
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    Currently not implemented. Recommended limits:
    - Public endpoints: 100 requests/minute
    - Admin endpoints: 1000 requests/minute
    
    ## Error Responses
    All endpoints return consistent error responses:
    ```json
    {
      "error": "Error message description",
      "code": "ERROR_CODE"
    }
    ```
    
  version: 1.0.0
  contact:
    name: API Support
    url: https://schoolswellbeing.com.au
  license:
    name: Proprietary
servers:
  - url: https://schoolswellbeing.com.au/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase authentication
...
paths:
...
  /admin/homepage-global-settings:
    get:
      tags:
        - Admin - Settings
      summary: Get global homepage settings
      operationId: getGlobalSettings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Global settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  global_colors:
                    type: object
                    properties:
                      primaryButton:
                        type: string
                      primaryButtonText:
                        type: string
                      secondaryButton:
                        type: string
                      secondaryButtonText:
                        type: string
                      heading:
                        type: string
                      subheading:
                        type: string
                      ctaBackground:
                        type: string
                      ctaText:
                        type: string
                      ctaPrimaryButton:
                        type: string

    patch:
      tags:
        - Admin - Settings
      summary: Update global homepage settings
      operationId: updateGlobalSettings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                global_colors:
                  type: object
      responses:
        '200':
          description: Settings updated successfully
...